refactor(tui): lift draw fns from main.rs into lib.rs (#1015) (#1029)
refactor(tui): extract rendering logic into view module (GH #1015)

- Move all per-view render functions from the `run<B>` closure in `main.rs`
  into a new `src/view.rs` module.
- Expose `pub fn draw(app, frame, theme, primer)` from `lib.rs` as the
  single rendering entry point.
- Reduce `main.rs` from 1167 to 401 LOC; add 807 LOC `view.rs`.
- Preserve existing behavior with all 146+ tests continuing to pass.
